move Subscriber to src/EventSubscriber
